+++ src/filter/PeakingEQFilter.cpp
@@ -28,7 +28,7 @@ dibiff::filter::Coefficients dibiff::filter::PeakingEQFilter::calculateCoefficie
     float w0 = 2.0f * M_PI * cutoff / sampleRate;
     float alpha = std::sin(w0) / (2.0f * qFactor);
     float b0 = 1.0f + (alpha * A);
-    float b1 = -2.0f * std::cosf(w0);
+    float b1 = -2.0f * std::cos(w0);
     float b2 = 1.0f - (alpha * A);
     float a0 = 1.0f + (alpha / A);
     float a2 = 1.0f - (alpha / A);
